+++ apps/cold-api/src/platform/modules/resources/actions/actions.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { CacheService, ColdCacheModule, JwtStrategy, MqttModule, PrismaModule } from '@coldpbc/nest';
+import { CacheService, ColdCacheModule, JwtStrategy, MqttModule } from '@coldpbc/nest';
 import { CategoryValidationModule } from '../categories/validation/category-validation.module';
 import { ActionTemplatesController } from './action-templates.controller';
 import { ActionTemplatesService } from './action-templates.service';
@@ -9,7 +9,7 @@ import { ActionsService } from './actions.service';
 import { SurveysModule } from '../surveys/surveys.module';
 
 @Module({
-  imports: [PrismaModule, ColdCacheModule, SurveysModule, CategoryValidationModule, MqttModule],
+  imports: [ColdCacheModule, SurveysModule, CategoryValidationModule, MqttModule],
   controllers: [ActionTemplatesController, ActionsController],
   providers: [ActionTemplatesService, ActionsService, JwtService, JwtStrategy, CacheService],
   exports: [ActionsService, ActionsService],