+++ apps/cold-ui/src/main.tsx
@@ -5,6 +5,7 @@ import { ColdContextProvider } from '@coldpbc/providers';
 import { datadogRum } from '@datadog/browser-rum';
 import { datadogLogs } from '@datadog/browser-logs';
 import { StyledEngineProvider, ThemeProvider } from '@mui/material';
+import { SWRConfig } from 'swr';
 import { muiTheme } from '@coldpbc/themes';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { LocalizationProvider } from '@mui/x-date-pickers';
@@ -45,6 +46,7 @@ datadogLogs.init({
 
 root.render(
   <StrictMode>
+    <SWRConfig>
     <ColdContextProvider
       auth0Options={{
         domain: domain,
@@ -59,12 +61,13 @@ root.render(
       }}
       launchDarklyClientSideId={launchDarklyClientSideId}>
       <LocalizationProvider dateAdapter={AdapterDateFns}>
-        <StyledEngineProvider injectFirst>
-          <ThemeProvider theme={muiTheme}>
-            <Home />
-          </ThemeProvider>
-        </StyledEngineProvider>
+      <StyledEngineProvider injectFirst>
+        <ThemeProvider theme={muiTheme}>
+          <Home />
+        </ThemeProvider>
+      </StyledEngineProvider>
       </LocalizationProvider>
     </ColdContextProvider>
+    </SWRConfig>
   </StrictMode>,
 );