+++ libs/nest/prisma/migrations/20240730141033_make_country_optional/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "organization_facilities" ALTER COLUMN "country" DROP NOT NULL,
+ALTER COLUMN "country" DROP DEFAULT;